fix(management): use 'portal' query param
When looking for APIs to subscribe from an application, we need to list the same APIs as if the user were on portal. To do this, we can use the query param 'portal' (which is only use for that case) to determine which APIs should be returned.

The application owner will be able to see only the published APIs that are public or of which he is a member.

Fixes gravitee-io/issues#5223