fix(api-review): check API has groups to avoid NPE
When looking for reviewers of the API, so they can be sent an email when a review is asked, we were looking into the groups of the API but without checking if this API had groups. So a NullPointerException happened when an API did not have any group.

Fixes gravitee-io/issues#5568