fix: add scopes in all BridgeCommand targets
An environment can be linked to both an APIM and AM installation on cockpit.
So to be able to send a command to the correct app cockpit need to know if it's targeting APIM or AM.
To avoid any issue we need to set a target scope to all BridgeCommand.